fix: 告警视图列表使用滚动加载--bug=155692464 # Reviewed, transaction id: 77846
