+++ bkmonitor/webpack/src/trace/pages/failure/failure-view/failure-view.scss
@@ -1,163 +1,171 @@
-.failure-view {
+.failure-view-wrapper {
   width: 100%;
   height: 100%;
-  padding: 16px;
+  padding: 16px 0px 16px 20px;
   background: #fff;
-  .metrics-collapse .correlation-metrics-collapse .correlation-metrics-collapse-head {
-    height: 32px;
-  }
-  @keyframes select {
-    0% {
-      background: #fafbfd;
-      border-color: #3a84ff;
-    }
-
-    35% {
-      background: #fafbfd;
-      border-color: #dcdee5;
-    }
-
-    70% {
-      background: #fafbfd;
-      border-color: #3a84ff;
-    }
-
-    100% {
-      background: white;
-      border-color: #dcdee5;
-    }
-  }
-
-  .failure-view-empty {
-    height: 400px;
-    margin-top: 10%;
-  }
-
-  .wm {
-    top: 12px;
-  }
-
-  .title-main {
-    font-size: 14px;
-    line-height: 40px;
-    color: #313238;
-
-    .title-num {
-      margin-left: 5px;
-      font-size: 12px;
-      color: #979ba5;
-    }
-  }
-
-  .dashboard-panel {
+  .failure-view {
+    width: 100%;
     height: 100%;
-  }
+    padding-right: 16px;
+    overflow: auto;
 
-  .correlation-metric-nav-wrap-bg {
-    position: absolute;
-    z-index: 1;
-    width: 259px;
-    height: calc(100% - 28px);
-    margin-right: 16px;
-    background: #fafbfd;
-    border-right: 1px solid #eaebf0;
-    border-radius: 2px 2px 0 0;
-  }
-
-  .correlation-metric-nav-wrap {
-    z-index: 9;
-    border-radius: 2px 2px 0 0;
-    transform: translateZ(0);
-  }
-
-  .correlation-metric-wrap {
-    position: relative;
-    display: flex;
-    flex-direction: row;
-    min-height: 260px;
-    padding-top: 22px;
-
-    .correlation-metric-fixed {
-      position: fixed;
-      z-index: 999;
+    .metrics-collapse .correlation-metrics-collapse .correlation-metrics-collapse-head {
+      height: 32px;
     }
+    @keyframes select {
+      0% {
+        background: #fafbfd;
+        border-color: #3a84ff;
+      }
 
-    .correlation-metric-fixed-padding {
-      padding-left: 275px;
-    }
+      35% {
+        background: #fafbfd;
+        border-color: #dcdee5;
+      }
 
-    .scroll-ing {
-      pointer-events: none;
-    }
+      70% {
+        background: #fafbfd;
+        border-color: #3a84ff;
+      }
 
-    .correlation-metrics-collapse-content {
-      margin-top: 12px;
-    }
-
-    .correlation-metric-panels {
-      flex: 1;
-      width: calc(100% - 275px);
-    }
-
-    .add-more {
-      display: inline-block;
-      width: 100%;
-      margin-top: 2px;
-      font-size: 14px;
-      line-height: 24px;
-      color: #3a84ff;
-      text-align: center;
-      cursor: pointer;
-      border-radius: 2px;
-
-      &:hover {
-        background: #f0f5ff;
+      100% {
+        background: white;
+        border-color: #dcdee5;
       }
     }
-  }
 
-  .metrics-err {
-    align-items: center;
-    padding-top: 0px;
-  }
-
-  .aiops-metrics-view-show {
-    display: block;
-  }
-
-  .aiops-metrics-view-hide {
-    display: none;
-  }
-
-  .bk-table-empty-text {
-    width: 100%;
-  }
-
-  .dashboard-panel {
-    width: 100%;
-  }
-
-  .aiops-dimension-panels {
-    padding-top: 40px;
-    margin-top: -40px;
-  }
-
-  .select {
-    animation: select 1s;
-    animation-delay: 0.5s;
-
-    .chart-wrapper,
-    .aiops-dimension-lint {
-      background: inherit;
+    .failure-view-empty {
+      height: 400px;
+      margin-top: 10%;
     }
-  }
 
-  .flex-dashboard {
-    padding-left: 0px;
+    .wm {
+      top: 12px;
+    }
 
-    .flex-dashboard-item {
-      border: 1px solid #dcdee5;
-      border-radius: 2px;
+    .title-main {
+      font-size: 14px;
+      line-height: 40px;
+      color: #313238;
+
+      .title-num {
+        margin-left: 5px;
+        font-size: 12px;
+        color: #979ba5;
+      }
+    }
+
+    .dashboard-panel {
+      height: auto;
+      overflow: visible;
+    }
+
+    .correlation-metric-nav-wrap-bg {
+      position: absolute;
+      z-index: 1;
+      width: 259px;
+      height: calc(100% - 28px);
+      margin-right: 16px;
+      background: #fafbfd;
+      border-right: 1px solid #eaebf0;
+      border-radius: 2px 2px 0 0;
+    }
+
+    .correlation-metric-nav-wrap {
+      z-index: 9;
+      border-radius: 2px 2px 0 0;
+      transform: translateZ(0);
+    }
+
+    .correlation-metric-wrap {
+      position: relative;
+      display: flex;
+      flex-direction: row;
+      min-height: 260px;
+      padding-top: 22px;
+
+      .correlation-metric-fixed {
+        position: fixed;
+        z-index: 999;
+      }
+
+      .correlation-metric-fixed-padding {
+        padding-left: 275px;
+      }
+
+      .scroll-ing {
+        pointer-events: none;
+      }
+
+      .correlation-metrics-collapse-content {
+        margin-top: 12px;
+      }
+
+      .correlation-metric-panels {
+        flex: 1;
+        width: calc(100% - 275px);
+      }
+
+      .add-more {
+        display: inline-block;
+        width: 100%;
+        margin-top: 2px;
+        font-size: 14px;
+        line-height: 24px;
+        color: #3a84ff;
+        text-align: center;
+        cursor: pointer;
+        border-radius: 2px;
+
+        &:hover {
+          background: #f0f5ff;
+        }
+      }
+    }
+
+    .metrics-err {
+      align-items: center;
+      padding-top: 0px;
+    }
+
+    .aiops-metrics-view-show {
+      display: block;
+    }
+
+    .aiops-metrics-view-hide {
+      display: none;
+    }
+
+    .bk-table-empty-text {
+      width: 100%;
+    }
+
+    .dashboard-panel {
+      width: 100%;
+    }
+
+    .aiops-dimension-panels {
+      padding-top: 40px;
+      margin-top: -40px;
+    }
+
+    .select {
+      animation: select 1s;
+      animation-delay: 0.5s;
+
+      .chart-wrapper,
+      .aiops-dimension-lint {
+        background: inherit;
+      }
+    }
+
+    .flex-dashboard {
+      padding-left: 0px;
+
+      .flex-dashboard-item {
+        border: 1px solid #dcdee5;
+        border-radius: 2px;
+      }
     }
   }
 }
