+++ bkmonitor/webpack/src/trace/plugins/components/dashboard-panel.scss
@@ -3,10 +3,8 @@
   // overflow: visible;
   flex: 1;
   width: 100%;
-  height: calc(100vh - 230px);
-  // max-height: 300px;
-  overflow: auto;
-  overflow-x: hidden;
+  height: auto;
+  overflow: visible;
 
   .vue-grid-layout {
     background: #f5f7fa;
@@ -122,6 +120,7 @@
   .flex-dashboard {
     display: flex;
     flex-wrap: wrap;
+    align-content: flex-start;
 
     &-row {
       display: flex;
@@ -149,5 +148,62 @@
         max-width: 100% !important;
       }
     }
+  }
+
+  .lazy-chart-placeholder {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    height: 100%;
+    background: #fafbfd;
+    border-radius: 2px;
+
+    .lazy-chart-loading {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+
+      .dot-loading {
+        display: flex;
+        gap: 4px;
+        align-items: center;
+
+        span {
+          display: inline-block;
+          width: 6px;
+          height: 6px;
+          background: #c4c6cc;
+          border-radius: 50%;
+          animation: dot-loading-bounce 1.4s ease-in-out infinite both;
+
+          &:nth-child(1) {
+            animation-delay: -0.32s;
+          }
+
+          &:nth-child(2) {
+            animation-delay: -0.16s;
+          }
+
+          &:nth-child(3) {
+            animation-delay: 0s;
+          }
+        }
+      }
+    }
+  }
+}
+
+@keyframes dot-loading-bounce {
+  0%,
+  80%,
+  100% {
+    opacity: 0.4;
+    transform: scale(0.8);
+  }
+
+  40% {
+    opacity: 1;
+    transform: scale(1);
   }
 }
